SAAS-320 Refactoring
